requirements.txt
autopep8==1.5.4
bitarray==1.5.3
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
pycodestyle==2.6.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
rope==0.17.0
toml==0.10.1
